+++ app/react/V2/Routes/Entity/Components/relationships/index.ts
@@ -1,7 +1,6 @@
 export {
   RelationshipsPanel,
   RelationshipRow,
-  CreateReference,
   CreateRelationshipModal,
   useRelationships,
   useRelationshipsActions,
