+++ app/react/V2/Routes/Entity/Components/relationships/panel/index.ts
@@ -1,6 +1,5 @@
 export { RelationshipsPanel } from './RelationshipsPanel.js';
 export { RelationshipRow } from '../rows/RelationshipRow.js';
-export { CreateReference } from '../create-reference/CreateReference.js';
 export { CreateRelationshipModal } from '../create-reference/CreateRelationshipModal.js';
 export {
   useRelationships,
